requirements.txt
xarray==2023.7.0
netCDF4<=1.6.4, <2.0.0
numpy<=1.25.2, <2.0.0
paho-mqtt<=1.6.1, <2.0.0
jsonschema<=4.19.0, <5.0.0
